+++ src/cpp/tardigrade-hydraLinearViscoelasticity.cpp
@@ -20,7 +20,9 @@ namespace tardigradeHydra{
              * 
              * \param &parameters: The paramter vector. Assumed to be a vector of length 2 which defines lambda and mu.
              */
-    
+   
+            const unsigned int *dim = hydra->getDimension( );
+ 
             if ( parameters.size( ) < 4 ){
     
                 ERROR_TOOLS_CATCH( throw std::runtime_error( "Parameter vector is expected to have a length of at least 4 but has a length of " + std::to_string( parameters.size( ) ) ) );
@@ -30,6 +32,19 @@ namespace tardigradeHydra{
             setNumVolumetricViscousTerms( ( unsigned int )( parameters[ 0 ] + 0.5 ) );
 
             setNumIsochoricViscousTerms( ( unsigned int )( parameters[ 1 ] + 0.5 ) );
+
+            setNumStateVariables( *getNumVolumetricViscousTerms( ) + ( *dim ) * ( *dim ) * ( *getNumIsochoricViscousTerms( ) ) );
+
+            if ( *getNumStateVariables( ) != ( *getViscoelasticISVUpperIndex( ) - *getViscoelasticISVLowerIndex( ) ) ){
+
+                std::string message = "The number of state variables required by the parameterization is not equal to the number of state variables indicated by the ISV bounds\n";
+                message            += "   required # ISVs: " + std::to_string( *getNumStateVariables( ) ) + "\n";
+                message            += "   ISV Lower Bound: " + std::to_string( *getViscoelasticISVLowerIndex( ) ) + "\n";
+                message            += "   ISV UPper Bound: " + std::to_string( *getViscoelasticISVLowerIndex( ) ) + "\n";
+
+                ERROR_TOOLS_CATCH( throw std::runtime_error( message ) );
+
+            }
 
             setKinf( parameters[ 2 ] );
 
@@ -174,17 +189,51 @@ namespace tardigradeHydra{
              * Decompose the elastic deformation into volumetric and isochoric parts
              */
 
-            const unsigned int *dim = hydra->getDimension( );
-
             floatType Je;
+
             floatVector Fehat;
 
-            ERROR_TOOLS_CATCH( Je = vectorTools::determinant( ( *hydra->getConfigurations( ) )[ 0 ], ( *dim ), ( *dim ) ) );
+            ERROR_TOOLS_CATCH( decomposeDeformation( ( *hydra->getConfigurations( ) )[ 0 ], Je, Fehat ) );
 
-            ERROR_TOOLS_CATCH( Fehat = ( *hydra->getConfigurations( ) )[ 0 ] / std::pow( Je, 1./3 ) );
             setJe( Je );
 
             setFehat( Fehat );
+
+        }
+
+        void residual::decomposePreviousElasticDeformation( ){
+            /*!
+             * Decompose the previous elastic deformation into volumetric and isochoric parts
+             */
+
+            floatType previousJe;
+
+            floatVector previousFehat;
+
+            ERROR_TOOLS_CATCH( decomposeDeformation( ( *hydra->getPreviousConfigurations( ) )[ 0 ], previousJe, previousFehat ) );
+
+            setPreviousJe( previousJe );
+
+            setPreviousFehat( previousFehat );
+
+        }
+
+        void residual::decomposeDeformation( const floatVector &F, floatType &J, floatVector &Fhat ){
+            /*!
+             * Decompose a deformation into volumetric and isochoric parts where
+             * 
+             * \f$\hat{\bf{F}} = J^{-\frac{1}{3}} \bf{F} \f$
+             * 
+             * \param &F: The incoming deformation gradient
+             * \param &J: The jacobian of deformation
+             * \param &Fhat: The isochoric part of the deformation gradient
+             */
+
+            const unsigned int *dim = hydra->getDimension( );
+
+            ERROR_TOOLS_CATCH( J = vectorTools::determinant( F, ( *dim ), ( *dim ) ) );
+
+            ERROR_TOOLS_CATCH( Fhat = F / std::pow( J, 1./3 ) );
 
         }
 
@@ -218,6 +267,34 @@ namespace tardigradeHydra{
 
         }
 
+        void residual::setPreviousJe( const floatType &previousJe ){
+            /*!
+             * Set the previous elastic Jacobian of deformation
+             * 
+             * \param &previousJe: The previous elastic Jacobian of deformation
+             */
+
+            _previousJe.second = previousJe;
+
+            _previousJe.first = true;
+
+        }
+
+        void residual::setPreviousFehat( const floatVector &previousFehat ){
+            /*!
+             * Set the previous isochoric part of the elastic deformation gradient
+             * 
+             * \param &previousFehat: The previous isochoric part of the elastic deformation gradient
+             */
+
+            _previousFehat.second = previousFehat;
+
+            _previousFehat.first = true;
+
+            addIterationData( &_previousFehat );
+
+        }
+
         const floatType* residual::getJe( ){
             /*!
              * Get the elastic Jacobian of deformation
@@ -245,6 +322,36 @@ namespace tardigradeHydra{
             }
 
             return &_Fehat.second;
+
+        }
+
+        const floatType* residual::getPreviousJe( ){
+            /*!
+             * Get the previous elastic Jacobian of deformation
+             */
+
+            if ( !_previousJe.first ){
+
+                ERROR_TOOLS_CATCH( decomposePreviousElasticDeformation( ) );
+
+            }
+
+            return &_previousJe.second;
+
+        }
+
+        const floatVector* residual::getPreviousFehat( ){
+            /*!
+             * Get the previous isochoric part of the elastic deformation gradient
+             */
+
+            if ( !_previousFehat.first ){
+
+                ERROR_TOOLS_CATCH( decomposePreviousElasticDeformation( ) );
+
+            }
+
+            return &_previousFehat.second;
 
         }
 
@@ -342,10 +449,111 @@ namespace tardigradeHydra{
 
         }
 
+        void residual::decomposeStateVariableVector( floatVector &volumetricStateVariables,
+                                                     floatVector &isochoricStateVariables ){
+            /*!
+             * Decompose the state variable vector into parameters associated with the
+             * volumetric and isochoric viscoelasticity
+             */
+
+            const unsigned int *dim = hydra->getDimension( );
+
+            unsigned int lb = *getViscoelasticISVLowerIndex( );
+            unsigned int ub = lb + *getNumVolumetricViscousTerms( );
+
+            volumetricStateVariables = floatVector( hydra->getAdditionalStateVariables( )->begin( ) + lb,
+                                                    hydra->getAdditionalStateVariables( )->begin( ) + ub );
+
+            lb = ub;
+
+            ub = lb + ( *dim ) * ( *dim ) * *getNumIsochoricViscousTerms( );
+
+            isochoricStateVariables = floatVector( hydra->getAdditionalStateVariables( )->begin( ) + lb,
+                                                   hydra->getAdditionalStateVariables( )->begin( ) + ub );
+
+        }
+
+        void residual::setNumStateVariables( const unsigned int numStateVariables ){
+            /*!
+             * Set the number of state variables
+             * 
+             * \param &numStateVariables: The number of state variables required
+             */
+
+            _numStateVariables = numStateVariables;
+
+        }
+
         void residual::setPK2Stress( ){
             /*!
              * Set the second Piola-Kirchhoff stress
              */
+
+            // Initialize required values
+
+            floatType time = *hydra->getTime( );
+
+            // Compute the strain measures
+
+            floatVector volumetricStrain = { ( *getJe( ) - 1 ) };
+
+            floatVector previousVolumetricStrain = { ( *getPreviousJe( ) - 1 ) };
+
+            floatVector isochoricStrain, previousIsochoricStrain;
+            ERROR_TOOLS_CATCH_NODE_POINTER( constitutiveTools::computeGreenLagrangeStrain( *getFehat( ), isochoricStrain ) );
+
+            ERROR_TOOLS_CATCH_NODE_POINTER( constitutiveTools::computeGreenLagrangeStrain( *getPreviousFehat( ), previousIsochoricStrain ) );
+
+            floatType previousTime = time - *hydra->getDeltaTime( );
+
+            // Get the previous state variable values
+
+            floatVector previousVolumetricStateVariables;
+
+            floatVector previousIsochoricStateVariables;
+
+            ERROR_TOOLS_CATCH( decomposeStateVariableVector( previousVolumetricStateVariables,
+                                                             previousIsochoricStateVariables ) );
+
+            floatVector PK2MeanStress;
+
+            floatVector deltaPK2MeanStress;
+
+            floatVector currentVolumetricStateVariables;
+
+            floatVector PK2IsochoricStress;
+
+            floatVector deltaPK2IsochoricStress;
+
+            floatVector currentIsochoricStateVariables;
+
+//            // Compute the viscous mean stress
+//
+//            ERROR_TOOLS_CATCH_NODE_POINTER( stressTools::linearViscoelasticity( *hydra->getTime( ), volumetricStrain,
+//                                                                                previousTime, previousVolumetricStrain,
+//                                                                                *getCurrentVolumetricRateModifier( ),
+//                                                                                *getPreviousVolumetricRateModifier( ),
+//                                                                                *getPreviousVolumetricStateVariables( ),
+//                                                                                *getVolumetricParameters( ),
+//                                                                                *getStressIntegrationAlpha( ), deltaPK2MeanStress,
+//                                                                                PK2MeanStress, currentVolumetricStateVariables ) );
+//
+//            // Compute the viscous isochoric stress
+//            ERROR_TOOLS_CATCH_NODE_POINTER( stressTools::linearViscoelasticity( *hydra->getTime( ), isochoricStrain,
+//                                                                                previousTime, previousIsochoricStrain,
+//                                                                                *getCurrentIsochoricRateModifier( ),
+//                                                                                *getPreviousIsochoricRateModifier( ),
+//                                                                                *getPreviousIsochoricStateVariables( ),
+//                                                                                *getIsochoricParameters( ),
+//                                                                                *getStressIntegrationAlpha( ), deltaPK2IsochoricStress,
+//                                                                                PK2IsochoricStress, currentIsochoricStateVariables ) );
+//
+//        floatVector eye( ( *dim ) * ( *dim ), 0 );
+//        vectorTools::eye( eye );
+//
+//        PK2Stress = PK2IsochoricStress + PK2MeanStress[ 0 ] * eye;
+//
+//        setPK2Stress( PK2Stress );
 
         }