+++ docs/sphinx/changelog.rst
@@ -49,3 +49,5 @@ Internal Changes
 - Changed the convergence_error type to use standard strings (:merge:`18`). By `Nathan Miller`_.
 - Changed the version extraction script (:merge:`19`). By `Nathan Miller`_.
 - Added elastic deformation gradient decomposition to linear viscoelasticity (:merge:`20`). By `Nathan Miller`_.
+- Generalized the decomposition of the current elastic deformation gradient to current and previous (:merge:`20`). By `Nathan Miller`_.
+- Added the decomposition of the additional state variable vector into volumetric and isochoric parts (:merge:`20`). By `Nathan Miller`_.